MacroAtomInfo.cpp file
